pyproject.toml
[tool.poetry]
name = "pdfding"
version = "0.1.0"
description = ""
authors = ["mrmn"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "==5.0.6"
django-allauth = {extras = ["socialaccount"], version = "==0.63.3"}
django-cleanup = "==8.1.0"
django-htmx = "==1.17.3"

[tool.poetry.group.dev.dependencies]
black = "==24.4.2"
flake8 = "==7.1.0"
coverage = "==7.5.4"

[tool.black]
line-length = 120
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[flake8]
max-line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
